All of the methods of this class are synchronized. When you open a lot of RrdB at the same time, it slows down the thing a lot.
Here is a patch to reduce locking I hope. It use a ConcurrentHashMap  
from Java 5, and if you open an already open store, there is no  
locking at all, except for reference counting.

I do not provide a method to change the default pool at runtime.
One can have a look here, from DbPool class, to change it:

	public static synchronized RrdDbPool getInstance() {
		if(RrdDbPoolSingletonHolder.instance == null) {
			RrdDbPoolSingletonHolder.instance = new RrdDbPoolNew();
		}
		return RrdDbPoolSingletonHolder.instance;
	}

The new code is backward and upward compatible with old one.

There is just a little trick : the capacity can't be changed if the  
pool is in use, it throws an exception if it happens.

git-svn-id: https://rrd4j.dev.java.net/svn/rrd4j/trunk/rrd4j@60 85f5bce8-2210-0410-b59b-f26aa68bfb99